requirements.txt
alpaca-py>=0.30.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
pyarrow>=15.0.0
lxml>=5.0.0
requests>=2.31.0
yfinance>=0.2.40
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
# Phase 7.5 Dashboard
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
